Add files via upload
OK, so all I did here was create a character class, modify drawCharacter() to handle a change of direction, and made a collision method in the Maze class. You should call the collision method as well as the respective movement when you create things to handle user input. All the code is commented but if you have any questions just ask. NOTE: I have not been able to test checkCollision() as of yet because of the fact that the character can't move yet, so it may not work.